heartrate: datapages: refector to new data api
With the new api we use common getters and setters for data which allows
us to extract common functions out of both the TX and RX functions. The
downside is virtual inheritance is needed to do the lookup. This also
reduces the amount of testing needed since only 1 set of get functions
needs testing now.