heartrate: refactor datapages
with core datapages i realized splitting the datapges into three files
was borderline obnoxious, so lets collapse everything back down. This is
step 1 of 2 of the refactor this profile. Second step is to trim down
defines by moving them where needed which will be done in the follwing
commit. This whole changewill lead to an overall text and code
reduction.