heartrate: move defines into datapages
now that datapages are consolidated we can stop writing mile long
defines, things are private as they are no longer in a header as well.
Also new diff is jsut sub -50 on line count. Added a couple todos but I
will be doing this refactor on the other profiles as well this year.